Add pedagogical notes about interpolation
I added a parenthetical about interpolation and another sentence which allows
them to dig deeper into what interpolation is if they are interested.

Missing data in and of itself is fairly complex (here it isn't MAR or MCAR either),
and the language of 'missing data' was absent before this; I think it's important
to frame this as a missing data problem---might need to provide scaffolds to help students see:

Data collected by decades ->
Want to predict survivability of any person in any given year ->
Need to "fill up the gaps" in the data ->
"Missing data" problem ->
What techniques are there to help with 'filling in' missing data ->

I think 'imputation' is a more common term and technique in intro/intermediate stats courses for missing data
problems; where the mean/median of the variable is substituted in (doesn't require linearity).

Wonder if talking about the linearity assumption of interpolation here is important or too in the weeds?